requirements.txt
python-dotenv==1.0.1
telethon==1.34.0
psycopg2-binary==2.9.9
dbt-postgres==1.7.7
ultralytics==8.2.20
fastapi==0.111.0
uvicorn==0.30.1
dagster==1.7.10
dagster-webserver==1.7.10
docker==7.0.0
